+++ smart_irrigation_system/node/interface/irrigation_cli.py
@@ -9,6 +9,7 @@ import time, logging, datetime
 from typing import Optional
 
 from smart_irrigation_system.node.core.irrigation_controller import IrrigationController
+from smart_irrigation_system.node.core.controller.controller_core import ControllerCore
 from smart_irrigation_system.node.utils.logger import get_dashboard_log_handler
 from smart_irrigation_system.__version__ import __version__ as version
 from smart_irrigation_system.node.core.enums import ControllerState
@@ -18,7 +19,7 @@ from smart_irrigation_system.node.core.status_models import CircuitSnapshot
 from smart_irrigation_system.node.core.enums import SnapshotCircuitState, IrrigationOutcome
 
 class IrrigationCLI:
-    def __init__(self, controller: IrrigationController, refresh_interval_idle=1, refresh_interval_active=0.1,
+    def __init__(self, controller: ControllerCore, refresh_interval_idle=1, refresh_interval_active=0.1,
                 max_logs=20, sleep_timeout=300):
         self.controller = controller
         self.refresh_interval_idle = refresh_interval_idle